requirements.txt
streamlit = 1.40.1
pandas = 2.2.2
matplotlib-base = 3.9.2           
matplotlib-inline = 0.1.6
plotly = 5.24.1
urllib3 = 2.2.3